new service for auth functions
- function for the auth endpoints are moved to a separate class, an auth
"service". This is not a RPC service.
